Update test to verify method calls. Remove throwaway tests
